init tag views module:
Due to the complexity of 3 tag CRUD views(totla of 15 views), I
seperated it depending on each models. Alcoho, Food and Place is now in
different files.
