+++ sooljottagrae/apis/views/tag/__init__.py
@@ -1,0 +1,3 @@
+from .alcohol import *
+from .food import *
+from .place import *
